Remove BraveSiteSettingsCounter, it's no longer necessary
With the migration away from ResourceIds, the BraveSiteSettingsCounter
class is no longer needed, since upstream's SiteSettingsCounter will be
good enough now that we don't need to perform a "subsearch" for the
ContentSettingsType::PLUGINS type, to consider Brave-specific cases.
￼
In this new worls where we have Brave-specific ContentSettingsTypes,
SiteSettingsCounter::Count() will suffice to account for Brave's own
ContentSettingsTypes, so there's no need to have a subclass to deal
with them now, therefore we have to remove it (otherwise the count
won't be correct, at least in its current form, since the same type
would be counted twice).